Rewrite the documentation's configuration and example usage.
The previous sample documentation in `index.rst` has been moved to
`csharpdomain-example.rst`, and updated to include the additional C# language
support (e.g. indexers). The actual `index.rst` is currently just a table of
contents, but could be expanded.

The `conf.py` configuration file has been adapted to use the newest options of
the latest `sphinx` - although the `nitpick_includes` mostly carried over.

Most frustratingly, none of the usual `sys.path` modifications enabled Python to
find the package `sphinxcontrib.csharpdomain`, because the `sphinxcontrib`
namespace package is already in `sys.modules` at the beginning of `conf.py`, due
to the `.pth` configuration files in `site-packages` that Python runs on startup
(unless started with `-S`).

Therefore, if sphinxcontrib.csharpdomain is importable (from a PyPI-based
`pip install` or even a local `pip install --editable
path/to/sphinxcontrib-csharpdomain`), everything will be fine. However, if not,
it's necessary to somehow tell `sphinxcontrib` to reevaluate where it looks for
subpackages. The best solution I could come up with is commented out at the head
of the file, but it's hacky at best, and it feels like there must be a better
way. I found [1] extremely useful in understanding namespace packages, so I'm
saving it here for future reference.

The new `Makefile` and `make.bat` come directly from `sphinx-quickstart`'s
generated code.

[1]: https://packaging.python.org/guides/packaging-namespace-packages/